Refactor startheremenu.tsx and fractalviewer.tsx into smaller modules
